Fix box price editing: remove .int() constraint from Zod schema
The price field is Decimal(10,2) in the DB but the API validation required an integer, rejecting any decimal price. Also ensure the GET endpoint returns price as a number instead of a Decimal string.

Made-with: Cursor

+++ src/app/api/admin/boxes/route.ts
@@ -8,7 +8,7 @@ const boxSchema = z.object({
   name: z.string().min(1),
   description: z.string().min(1),
   imageUrl: z.string().min(1), // Allow any string, not just URLs (some APIs return relative paths)
-  price: z.number().int().positive(),
+  price: z.number().positive(),
   cardsPerPack: z.number().int().positive(),
   games: z.array(z.enum(['MAGIC_THE_GATHERING', 'ONE_PIECE', 'POKEMON', 'LORCANA', 'YUGIOH', 'FLESH_AND_BLOOD'])).min(1),
 });
